(1)将采样的原始数据类型改为Float (2)加入滤波功能

+++ src/com/lon/outsidemonitor/SignalDetailActivity.java
@@ -54,7 +54,7 @@ public class SignalDetailActivity extends Activity {
 					signalDetailView.refreshRawData(null);
 					signalDetailView.refreshSpectrumData(null);
 				} else {
-					short[] rawData = signal.getRawData();
+					float[] rawData = signal.getRawData();
 					float[] spectrumData = signal.getSpectrumData();
 					signalDetailView.refreshRawData(rawData);
 					signalDetailView.refreshSpectrumData(spectrumData);
